+++ src/components/forms/button.tsx
@@ -8,6 +8,7 @@ interface ButtonProps {
     px?: number;
     py?: number;
     onClick: Function;
+    disabled?: boolean;
 }
 
 export default function Button(props: ButtonProps) {
@@ -20,7 +21,7 @@ export default function Button(props: ButtonProps) {
     const classes = `px-${props.px || 4} py-${props.py || 2} rounded-md bg-${props.bg || 'white'} border border-${props.colour || 'blue'}-600 text-${props.colour || 'blue'}-600 text-${props.size}`
 
     return (
-        <button className={classes} onClick={handleClick}>
+        <button className={classes} onClick={handleClick} disabled={props.disabled}>
             {props.text}
         </button>
     )